refactor: split cli parser setup
